moved num_ref.h to src/common..., added another MR image data classes layer
