+++ src/common/include/sirf/common/num_ref.h
